+++ api/python/src/nanobind/extra/stl/wstring.h
@@ -9,7 +9,7 @@ NAMESPACE_BEGIN(detail)
 template <> struct type_caster<std::wstring> {
     NB_TYPE_CASTER(std::wstring, const_name("Union[str, bytes]"));
 
-    bool from_python(handle src, uint8_t, cleanup_list *) noexcept {
+    bool from_python(handle src, uint32_t, cleanup_list *) noexcept {
         Py_ssize_t size;
         const wchar_t *str = PyUnicode_AsWideCharString(src.ptr(), &size);
         if (str == nullptr) {